+++ src/Credfeto.Exceptions.SourceGenerator/IsExternalInit.cs
@@ -1,0 +1,8 @@
+// This file is needed to support C# record types when targeting netstandard2.0.
+// ReSharper disable All
+// ReSharper disable CheckNamespace
+namespace System.Runtime.CompilerServices;
+
+internal static class IsExternalInit
+{
+}